==> notebooks/2023-02-25_validation/2. Validate defence model.ipynb <==
from pathlib import Path
import numpy as np
import pandas as pd
from sklearn.metrics import roc_auc_score
from sklearn.model_selection import StratifiedKFold

import sys, os
from pathlib import Path

sys.path.insert(0, os.path.join(Path('.').resolve().parent.parent))

from fusionlib.model import predict # Функция, позволяет получить предсказание нейронки.
from fusionlib.check_budget import check_budget # функция проверки бюджета. Проверяйте допустимость решения до сабмита

clear_df = pd.read_csv('/Users/xxx/Documents/Programming/vtb-data-fusion-2023/vtb-data-fusion-2023-defence/data/validation_clear_transactions.csv')
clear_df.tail(5)

fraud_df = pd.read_csv('/Users/xxx/Documents/Programming/vtb-data-fusion-2023/vtb-data-fusion-2023-defence/data/validation_fraud_transactions.csv')
fraud_df.tail(5)

clear_df_path = '/Users/xxx/Documents/Programming/vtb-data-fusion-2023/vtb-data-fusion-2023-defence/data/validation_clear_transactions.csv'
fraud_df_path = '/Users/xxx/Documents/Programming/vtb-data-fusion-2023/vtb-data-fusion-2023-defence/data/validation_fraud_transactions.csv'
bins_path = "/Users/xxx/Documents/Programming/vtb-data-fusion-2023/vtb-data-fusion-2023-defence/models/nn_bins.pickle" # путь до файла с бинами после тренировки модели (nn_bins.pickle)
model_path = "/Users/xxx/Documents/Programming/vtb-data-fusion-2023/vtb-data-fusion-2023-defence/models/nn_weights.ckpt" # путь до файла с весами нейронной сети (nn_weights.ckpt)
quantiles_path = "/Users/xxx/Documents/Programming/vtb-data-fusion-2023/vtb-data-fusion-2023-defence/misc/quantiles.json" # путь до файла с квантилями для таргета (quantiles.pickle)
BUDGET = 10 # разрешенное количество изменений транзакций для каждого пользователя

%%time
# у нас нет разметки для тех транзакций, которые мы атакуем - но у нас есть модель.
# Давайте посчитаем вероятность того, что пользователь принадлежит к классу 1
result_clear = predict(clear_df_path, bins_path, model_path, random_seed=20230206)
result_clear = result_clear.rename(columns={'target': 'clear_proba'})

%%time
result_fraud = predict(fraud_df_path, bins_path, model_path, random_seed=20230206)
result_fraud = result_fraud.rename(columns={'target': 'fraud_proba'})

targets_path = '/Users/xxx/Documents/Programming/vtb-data-fusion-2023/data/target_finetune.csv'
targets = pd.read_csv(targets_path)

result = result_clear.merge(result_fraud, how='inner', left_on='user_id', right_on='user_id')
result = result.merge(targets, how='left', left_on='user_id', right_on='user_id')
result.head()

roc_aucs = []
skf = StratifiedKFold(n_splits=5)
for i, (train_index, test_index) in enumerate(skf.split(result, result.target)):
    test = result.iloc[test_index]
    roc_auc_original = roc_auc_score(test.target, test.clear_proba)
    roc_auc_attacked = roc_auc_score(test.target, test.fraud_proba)
    roc_auc = 2 / (1 / roc_auc_original  + 1 / roc_auc_attacked)
    roc_aucs.append(roc_auc)

np.mean(roc_aucs)

def convert_different_inputs_to_df(method_input: (str | pd.DataFrame)) -> pd.DataFrame:
    if isinstance(method_input, str):
        object = pd.read_csv(method_input)
    elif isinstance(method_input, pd.DataFrame):
        object = method_input
    return object

def validate_results(result_clear: (str | pd.DataFrame), 
                   result_fraud: (str | pd.DataFrame), 
                   targets: (str | pd.DataFrame), n_splits=5) -> float:
    
    result_clear = convert_different_inputs_to_df(result_clear)
    result_fraud = convert_different_inputs_to_df(result_fraud)
    targets = convert_different_inputs_to_df(targets)
    result = result_clear.merge(result_fraud, how='inner', left_on='user_id', right_on='user_id')
    result = result.merge(targets, how='left', left_on='user_id', right_on='user_id')
    roc_aucs = []
    skf = StratifiedKFold(n_splits=n_splits, random_state=42, shuffle=True)
    for i, (_, test_index) in enumerate(skf.split(result, result.target)):
        test = result.iloc[test_index]
        roc_auc_original = roc_auc_score(test.target, test.clear_proba)
        roc_auc_attacked = roc_auc_score(test.target, test.fraud_proba)
        roc_auc = 2 / (1 / roc_auc_original  + 1 / roc_auc_attacked)
        print(f'Validation iter={i}, roc_auc={roc_auc}')
        roc_aucs.append(roc_auc)
    
    return np.mean(roc_aucs)

from fusionlib.validation import validate_results
mean_harm_roc_auc = validate_results(result_clear, result_fraud, targets, n_splits=5)
print('-------------------- result -------------------- ')
print(f'mean_harm_roc_auc={mean_harm_roc_auc}')

public_mean_harm_roc_auc = 0.700928
print(f'Разница лидерборда и локальной валидации: {public_mean_harm_roc_auc - mean_harm_roc_auc}')

# опционально добавил возможность использовать как pd.DataFrame, так и уже сохраненные csv
validate_results(result_clear, result_fraud, '/Users/xxx/Documents/Programming/vtb-data-fusion-2023/data/target_finetune.csv')

# Метод скоринга кастомных predict методов

Чтобы снизить порог входа для данной задачи, был разработан метод оценки модели защиты, который позволяет не углубляться в механику атаки. Для оценки качества очередной версии своей модели необходимо лишь передать predict метод в соответствующую функцию.

def validate_model(predict_method: object, clear_df_path: str, 
                   fraud_df_path: str, bins_path: str, model_path: str, targets_path: str, random_seed=20230206, n_splits=5) -> float:
    # загрузка выбранных методов
    targets = pd.read_csv(targets_path)
    # получение скоров
    result_clear = predict_method(clear_df_path, bins_path, model_path, random_seed=random_seed)
    result_fraud = predict_method(fraud_df_path, bins_path, model_path, random_seed=random_seed)

    result_clear = result_clear.rename(columns={'target': 'clear_proba'})
    result_fraud = result_fraud.rename(columns={'target': 'fraud_proba'})
    
    mean_harm_roc_auc = validate_results(result_clear, result_fraud, targets, n_splits=n_splits)
    return mean_harm_roc_auc

validate_model(predict, clear_df_path, fraud_df_path, bins_path=bins_path, model_path=model_path, targets_path=targets_path)

source_file = clear_df_path
df_transactions = (
        pd.read_csv(
            source_file,
            parse_dates=["transaction_dttm"],
            dtype={"user_id": int, "mcc_code": int, "currency_rk": int, "transaction_amt": float},
        )
        .dropna()
        .assign(
            hour=lambda x: x.transaction_dttm.dt.hour,
            day=lambda x: x.transaction_dttm.dt.dayofweek,
            month=lambda x: x.transaction_dttm.dt.month,
            number_day=lambda x: x.transaction_dttm.dt.day,
        )
    )

df_transactions
